+++ src/acceptance-test/java/org/zalando/nakadi/webservice/CursorsServiceAT.java
@@ -23,6 +23,7 @@ import org.zalando.nakadi.service.subscription.model.Partition;
 import org.zalando.nakadi.service.subscription.zk.NewZkSubscriptionClient;
 import org.zalando.nakadi.service.subscription.zk.SubscriptionClientFactory;
 import org.zalando.nakadi.service.timeline.TimelineService;
+import org.zalando.nakadi.util.UUIDGenerator;
 import org.zalando.nakadi.view.SubscriptionCursorWithoutToken;
 import org.zalando.nakadi.webservice.utils.ZookeeperTestUtils;
 
@@ -64,6 +65,7 @@ public class CursorsServiceAT extends BaseAT {
     private CursorConverter cursorConverter;
     private CursorsService cursorsService;
     private EventTypeRepository eventTypeRepository;
+    private UUIDGenerator uuidGenerator;
 
     private String etName;
     private String topic;
@@ -115,8 +117,10 @@ public class CursorsServiceAT extends BaseAT {
         final SubscriptionDbRepository subscriptionRepo = mock(SubscriptionDbRepository.class);
         when(subscriptionRepo.getSubscription(sid)).thenReturn(subscription);
         final SubscriptionClientFactory zkSubscriptionFactory = new SubscriptionClientFactory(zkHolder, MAPPER);
+        uuidGenerator = mock(UUIDGenerator.class);
+        when(uuidGenerator.isUUID(any())).thenReturn(true);
         cursorsService = new CursorsService(timelineService, subscriptionRepo, eventTypeRepository,
-                mock(NakadiSettings.class), zkSubscriptionFactory, cursorConverter);
+                mock(NakadiSettings.class), zkSubscriptionFactory, cursorConverter, uuidGenerator);
 
         // Register cursors in converter
         registerNakadiCursor(new NakadiCursor(createFakeTimeline(etName, topic), P1, NEW_OFFSET));
@@ -150,6 +154,13 @@ public class CursorsServiceAT extends BaseAT {
         } catch (final InvalidStreamIdException ignore) {
         }
         checkCurrentOffsetInZk(P1, OLD_OFFSET);
+    }
+
+    @Test(expected = InvalidStreamIdException.class)
+    public void shouldThrowInvalidStreamIdWhenStreamIdIsNotUUID() throws Exception {
+        when(uuidGenerator.isUUID(any())).thenReturn(false);
+        final String streamId = "/";
+        cursorsService.commitCursors(streamId, sid, testCursors);
     }
 
     @Test